requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
alembic==1.13.1

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client (for BuzzerBeater API)
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# XML parsing (BuzzerBeater API returns XML)
lxml==5.1.0

# Development
python-dotenv==1.0.0

# Deployment
greenlet==3.3.0
a2wsgi==1.10.0
